Removed the exogenous_change() code and placed it in the definition of the VacuumEnvironment instance.  Should this be a method that is passed to the function?
Created a NewVacuumEnvironment constructor that handles the individual configurations with an if config="config": elif: else: structure.  The three configurations currently implemented are 'center walls', 'full dirt', 'center walls w/ random dirt and fire'

Added code to set the seed in the main() function to provide repeatable random results.

Modified the left_click function on the GUI to print the performance of any agents in the reported cell.

Added a right_click function to the GUI to show what objects are being held by any agents in the reported cell.

Added some misc comments.